feat: Pauta register

+++ src/main/java/com/db/desafio/votacao/api/v1/modules/votacaoAssembleia/database/AssembleiaRepository.java
@@ -26,7 +26,7 @@ import com.db.desafio.votacao.api.v1.modules.votacaoAssembleia.database.models.A
 @Repository
 public interface AssembleiaRepository 
     extends
-        JpaRepository<Assembleia, Integer> 
+        JpaRepository<Assembleia, Long> 
 {
 
 }
